Implement Typical Sampling
PiperOrigin-RevId: 565071054
